JobPortal_WebApp/requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
azure-core==1.33.0
azure-storage-blob==12.25.1
beautifulsoup4==4.13.3
bitsandbytes==0.45.4
blinker==1.9.0
bs4==0.0.2
cachelib==0.13.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.10
crispy-bootstrap5==2025.4
cryptography==44.0.2
cut-cross-entropy==25.1.1
datasets==3.5.0
diffusers==0.32.2
dill==0.3.8
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.7
django-ckeditor==6.7.2
django-crispy-forms==2.4
django-debug-toolbar==5.1.0
django-heroku==0.3.1
django-js-asset==3.1.2
django-taggit==6.1.0
django-user-visit==2.0
docopt==0.6.2
docstring_parser==0.16
easyocr==1.7.2
filelock==3.16.1
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
frozenlist==1.5.0
fsspec==2024.12.0
google_search_results==2.4.2
greenlet==3.1.1
h11==0.14.0
hf_transfer==0.1.9
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
Js2Py==0.74
lazy_loader==0.4
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
msgspec==0.19.0
multidict==6.3.2
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.1
ollama==0.4.7
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
peft==0.15.1
pillow==11.1.0
pipwin==0.5.2
platformdirs==4.3.6
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
psycopg2==2.9.10
pyarrow==19.0.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
PyGetWindow==0.0.9
Pygments==2.19.1
pyjsparser==2.7.1
PyJWT==2.10.1
PyMsgBox==1.0.9
pyperclip==1.9.0
pypiwin32==223
PyPrind==2.11.3
PyRect==0.2.0
PyScreeze==1.0.1
pySmartDL==1.3.4
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.98
pytweening==1.2.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
sentencepiece==0.2.0
setuptools==78.1.0
shapely==2.1.0
shtab==1.7.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.2
SQLAlchemy==2.0.38
sqlparse==0.5.3
sympy==1.13.1
tabulate==0.9.0
tifffile==2025.3.30
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.50.3
triton-windows==3.2.0.post17
trl==0.15.2
typeguard==4.4.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tyro==0.9.18
tzdata==2025.2
tzlocal==5.3.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
unsloth==2025.3.19
unsloth_zoo==2025.3.17
urllib3==2.3.0
user-agents==2.2.0
virtualenv==20.28.0
virtualenvwrapper-win==1.2.7
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.9.0
WTForms==3.2.1
xformers==0.0.29.post3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
